+++ lib/hausaufgabenheft_logik/lib/src/data_source/homework_data_source.dart
@@ -6,7 +6,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
 import 'package:collection/collection.dart' show DeepCollectionEquality;
 
 abstract class HomeworkDataSource {
