+++ lib/hausaufgabenheft_logik/lib/src/lehrer/teacher_homework_view_factory.dart
@@ -6,7 +6,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:hausaufgabenheft_logik/src/models/homework/models_used_by_homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/models_used_by_homework.dart';
 import 'package:hausaufgabenheft_logik/src/views/color.dart';
 
 import '../../hausaufgabenheft_logik_lehrer.dart';
