+++ lib/hausaufgabenheft_logik/lib/src/models/homework_completion_status.dart
