+++ lib/hausaufgabenheft_logik/lib/src/models/models_used_by_homework.dart
@@ -8,7 +8,7 @@
 
 library models;
 
-export '../date.dart';
-export '../subject.dart';
-export '../title.dart';
+export 'date.dart';
+export 'subject.dart';
+export 'title.dart';
 export 'homework_completion_status.dart';
