+++ lib/hausaufgabenheft_logik/lib/src/open_homeworks/open_homework_view_bloc/events.dart
@@ -7,7 +7,7 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 import 'package:equatable/equatable.dart';
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
 import 'package:hausaufgabenheft_logik/src/open_homeworks/sort_and_subcategorization/sort/src/sort.dart';
 
 abstract class OpenHomeworkViewEvent extends Equatable {}
