+++ lib/hausaufgabenheft_logik/lib/src/open_homeworks/sort_and_subcategorization/sort/smallest_date_subject_and_title_sort.dart
@@ -6,7 +6,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import '../../../models/homework/homework.dart';
+import '../../../models/homework.dart';
 import '../../../models/date.dart';
 import 'src/sort.dart';
 import 'src/sort_with_operations.dart';
