+++ lib/hausaufgabenheft_logik/lib/src/open_homeworks/sort_and_subcategorization/sort_and_subcategorizer.dart
@@ -8,7 +8,7 @@
 
 import 'package:hausaufgabenheft_logik/src/models/homework_list.dart';
 import 'package:hausaufgabenheft_logik/src/open_homeworks/sort_and_subcategorization/sort/src/sort.dart';
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
 import 'package:hausaufgabenheft_logik/src/open_homeworks/views/homework_section_view.dart';
 
 import 'subcategorizer.dart';
