+++ lib/hausaufgabenheft_logik/lib/src/views/student_homework_view_factory.dart
@@ -6,8 +6,8 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
-import 'package:hausaufgabenheft_logik/src/models/homework/models_used_by_homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/models_used_by_homework.dart';
 import 'package:hausaufgabenheft_logik/src/views/color.dart';
 import 'package:hausaufgabenheft_logik/src/views/homework_view.dart';
