+++ lib/hausaufgabenheft_logik/test/create_homework_util.dart
@@ -7,8 +7,8 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 import 'package:common_domain_models/common_domain_models.dart';
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
-import 'package:hausaufgabenheft_logik/src/models/homework/models_used_by_homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/models_used_by_homework.dart';
 import 'package:hausaufgabenheft_logik/src/views/color.dart';
 import 'package:test_randomness/test_randomness.dart';
