+++ lib/hausaufgabenheft_logik/test/homework_list_test.dart
@@ -6,8 +6,8 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
-import 'package:hausaufgabenheft_logik/src/models/homework/models_used_by_homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/models_used_by_homework.dart';
 import 'package:hausaufgabenheft_logik/src/models/homework_list.dart';
 import 'package:hausaufgabenheft_logik/src/open_homeworks/sort_and_subcategorization/sort/smallest_date_subject_and_title_sort.dart';
 import 'package:hausaufgabenheft_logik/src/open_homeworks/sort_and_subcategorization/sort/subject_smallest_date_and_title_sort.dart';
