+++ lib/hausaufgabenheft_logik/test/in_memory_repo/realtime_completed_homework_loader.dart
@@ -6,7 +6,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:hausaufgabenheft_logik/src/models/homework/homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/homework.dart';
 
 abstract class RealtimeCompletedHomeworkLoader {
   Stream<List<HomeworkReadModel>> loadMostRecentHomeworks(
