+++ lib/hausaufgabenheft_logik/test/test_data/homeworks.dart
@@ -6,7 +6,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-import 'package:hausaufgabenheft_logik/src/models/homework/models_used_by_homework.dart';
+import 'package:hausaufgabenheft_logik/src/models/models_used_by_homework.dart';
 
 import '../create_homework_util.dart';
